Create Get-PassPhrase.ps1
This is a module that can be imported and used in your own scripts:
"Find users with weak passwords and set a strong passphrase!" 
"Generate new users with a strong passphrase.".
or others.

Please treat the generated secrets with caution, e.g. do not distribute in an unsafe manner such as e-mail!